Fishbook and CommandEffect Update
