+++ Fishoo/Assets/Prefabs/Command/BaseCommand.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 397f4700e50819c49b900292d9572f1f
-folderAsset: yes
-DefaultImporter:
+guid: 8b92c4d2dcddb9b4eb339c6b21ab8964
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
