+++ Fishoo/Assets/Prefabs/Command/CommandA.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: faab87364319dac48895e746908bb34d
+guid: 37a057046123e1a49b3cf6d832e987da
 PrefabImporter:
   externalObjects: {}
   userData:
