+++ Fishoo/Assets/Prefabs/Command/CommandB.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2683611989303c74fa52df3cdd04a5c0
+guid: d965a85782dc36b468c2467b55db1d99
 PrefabImporter:
   externalObjects: {}
   userData:
