+++ Fishoo/Assets/Prefabs/Command/CommandContainer.prefab
@@ -300,7 +300,7 @@ GameObject:
   - component: {fileID: 4858669746828714945}
   - component: {fileID: 3571153352457281984}
   m_Layer: 5
-  m_Name: Command
+  m_Name: CommandContainer
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -322,10 +322,10 @@ RectTransform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
-  m_AnchorMin: {x: 0.5, y: 0.5}
-  m_AnchorMax: {x: 0.5, y: 0.5}
+  m_AnchorMin: {x: 0, y: 0}
+  m_AnchorMax: {x: 1, y: 1}
   m_AnchoredPosition: {x: 0, y: 0}
-  m_SizeDelta: {x: 100, y: 100}
+  m_SizeDelta: {x: -598, y: -292}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &3571153352457281984
 MonoBehaviour:
