+++ Fishoo/Assets/Prefabs/Command/CommandDown.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 559032c26d4467c4081572714f94bce2
+guid: 86977cf324e8b6b4aac71da073d01461
 PrefabImporter:
   externalObjects: {}
   userData:
