+++ Fishoo/Assets/Prefabs/Command/CommandEraceEffect.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 07571d50defc2994b9d66438d8ff7128
-folderAsset: yes
-DefaultImporter:
+guid: ecb3fe4186640154e9a574b7a8149153
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
