+++ Fishoo/Assets/Prefabs/Command/CommandLeft.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8547017a7f563a04f96ba2d73b63f841
+guid: 2c59592f10a42ab438bcc54e741d17f3
 PrefabImporter:
   externalObjects: {}
   userData:
