+++ Fishoo/Assets/Prefabs/Command/CommandRight.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a56be007c9c4f3d4b9cdc801c3551e50
+guid: 5c4866c4da5f45340aac75fc979397aa
 PrefabImporter:
   externalObjects: {}
   userData:
