+++ Fishoo/Assets/Prefabs/Command/CommandUp.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6218e06e5c698554a951aa801fcf2dbc
+guid: 5f5dad8d707b0614f86d7cba4099a05c
 PrefabImporter:
   externalObjects: {}
   userData:
