+++ Fishoo/Assets/Prefabs/Effect/CommandEraceEffect.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: fe9655bde3f9b264893a240808a21ae6
-PrefabImporter:
+guid: f7f4145bcedcc744eac7baf36bd0e77a
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
