+++ Fishoo/Assets/Prefabs/Effect/CommandEraceFlareEffect.mat.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: eaf00d956fc52d5428d9fda452c69416
-folderAsset: yes
-DefaultImporter:
+guid: 1955e90444824ed4591b009e5acbd2b5
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
