+++ Fishoo/Assets/Prefabs/Fishing/CommandGenerator.prefab
@@ -44,30 +44,31 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   commandParent: {fileID: 0}
-  commandPrefab: {fileID: 3376987093915380326, guid: f4ff59cc081a0b045ae4d4f82b404a9f,
+  commandContainerPrefab: {fileID: 3376987093915380326, guid: f4ff59cc081a0b045ae4d4f82b404a9f,
     type: 3}
   commandsData:
   - key: 97
-    prefab: {fileID: 3218857649303722864, guid: faab87364319dac48895e746908bb34d,
+    prefab: {fileID: 9130246656005822197, guid: 37a057046123e1a49b3cf6d832e987da,
       type: 3}
   - key: 98
-    prefab: {fileID: 3218857649303722864, guid: 2683611989303c74fa52df3cdd04a5c0,
+    prefab: {fileID: 9130246656005822197, guid: d965a85782dc36b468c2467b55db1d99,
       type: 3}
   - key: 120
-    prefab: {fileID: 3218857649303722864, guid: 5d3607dea51815d4e8982ea23bc186be,
+    prefab: {fileID: 9130246656005822197, guid: 161fb8ea73a2ec940b885c84e282fa66,
       type: 3}
   - key: 121
-    prefab: {fileID: 3218857649303722864, guid: 8be7b175e4c59204a87e4b231676f62e,
+    prefab: {fileID: 9130246656005822197, guid: fe9655bde3f9b264893a240808a21ae6,
       type: 3}
   - key: 273
-    prefab: {fileID: 3218857649303722864, guid: 6218e06e5c698554a951aa801fcf2dbc,
+    prefab: {fileID: 9130246656005822197, guid: 5f5dad8d707b0614f86d7cba4099a05c,
       type: 3}
   - key: 274
-    prefab: {fileID: 3218857649303722864, guid: 559032c26d4467c4081572714f94bce2,
+    prefab: {fileID: 9130246656005822197, guid: 86977cf324e8b6b4aac71da073d01461,
       type: 3}
   - key: 275
-    prefab: {fileID: 3218857649303722864, guid: a56be007c9c4f3d4b9cdc801c3551e50,
+    prefab: {fileID: 9130246656005822197, guid: 5c4866c4da5f45340aac75fc979397aa,
       type: 3}
   - key: 276
-    prefab: {fileID: 3218857649303722864, guid: 8547017a7f563a04f96ba2d73b63f841,
+    prefab: {fileID: 9130246656005822197, guid: 2c59592f10a42ab438bcc54e741d17f3,
       type: 3}
+  targetFish: {fileID: 0}
