+++ Fishoo/Assets/Scene/CommandTest.unity
@@ -98,7 +98,7 @@ LightmapSettings:
     m_TrainingDataDestination: TrainingData
     m_LightProbeSampleCountMultiplier: 4
   m_LightingDataAsset: {fileID: 0}
-  m_UseShadowmask: 0
+  m_UseShadowmask: 1
 --- !u!196 &4
 NavMeshSettings:
   serializedVersion: 2
@@ -762,6 +762,11 @@ PrefabInstance:
     - target: {fileID: 916627375920575450, guid: c5331c5290144bd4a89e6432cd71a142,
         type: 3}
       propertyPath: commandParent
+      value: 
+      objectReference: {fileID: 1389284977}
+    - target: {fileID: 916627375920575450, guid: c5331c5290144bd4a89e6432cd71a142,
+        type: 3}
+      propertyPath: commandEffectParent
       value: 
       objectReference: {fileID: 1389284977}
     - target: {fileID: 916627375920575451, guid: c5331c5290144bd4a89e6432cd71a142,
