+++ Fishoo/Assets/Scripts/EffectScript.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1553fcf61c2f0f54d85093d330d63ccb
+guid: 829e8557f9b24154c9a8800b33aeb871
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
