+++ Fishoo/Assets/Scripts/Sprite.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d3a11b55ed191fb459906e576f623e66
+guid: 03a5367dddeeadc4db212859ab144782
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
