+++ Fishoo/Assets/Shader.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: fe9655bde3f9b264893a240808a21ae6
-PrefabImporter:
+guid: 9fa4c2ff00e1dfb4a9b94014ba082d46
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
