+++ Fishoo/Assets/Sprite.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: bb487d478760cb848b3681c94d28286a
+guid: d3a11b55ed191fb459906e576f623e66
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
