+++ Fishoo/Assets/Sprite/Player.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: fe9655bde3f9b264893a240808a21ae6
-PrefabImporter:
+guid: 76817bfe9fae9644084876a12baf8174
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
